Minimised dynamic allocations, can't completely stop potential memory leaks yet, because FreeRTOS-Plus-TCP itself isn't able to be completely static

+++ src/program/test.cpp
@@ -9,8 +9,8 @@ REGISTER_ROUTINE(point_mass_model, 10)
     auto acc = System::get_parameter<double>("acceleration");
 
     // perform a calculation with the parameters
-    pos = pos + vel * (double)(1 / 10.0);
-    vel = vel + acc * (double)(1 / 10.0);
+    pos = pos + vel * (1 / 10.0);
+    vel = vel + acc * (1 / 10.0);
 
     // update parameter 2 by name
     System::set_parameter("position", pos);
